// src/packages/h5p-server/types.ts
export interface ILibraryName {
    /**
     * The name used to identify the library (e.g. H5P.Example)
     */
    machineName: string;
    /**
     * The major version of the library (e.g. 1)
     */
    majorVersion: number;
    /**
     * The minor version of the library (e.g. 0)
     */
    minorVersion: number;
}
// ...
export interface IUser {
    /**
     * If true, the user can create content of content types that are set to "restricted".
     */
    canCreateRestricted: boolean;
    /**
     * If true, the user can install content types from the hub that are set the "recommended"
     * by the Hub.
     */
    canInstallRecommended: boolean;
    /**
     * If true, the user can generally install and update libraries. This includes Hub
     * content types that aren't set to "recommended" or uploading custom packages.
     */
    canUpdateAndInstallLibraries: boolean;
    /**
     * E-Mail address.
     */
    email: string;
    /**
     * An internal id used to check if user objects are identical.
     */
    id: string;
    /**
     * The full name of the user.
     */
    name: string;
    /**
     * Specifies type of user. Possible values other than 'local' are unknown as of this time.
     */
    type: 'local' | string;
}
// ...
export type ContentParameters = any;
// ...
export interface IPlayerModel {
    contentId: ContentParameters;
    argsId: IH5PPlayerArgs;
    // contentId: ContentParameters;
    dependencies: ILibraryName[];
    downloadPath: string;
    embedTypes: ('iframe' | 'div')[];
    integration: IIntegration;
    scripts: string[];
    styles: string[];
    translations: any;
    user: IUser;
}